% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOVDBIB.R, R/DBIB.R
\name{Aov_DBIB}
\alias{Aov_DBIB}
\title{Diseño de bloques incompletos balanceados}
\usage{
Aov_DBIB(respuesta, tratamiento, bloque, data)

Aov_DBIB(respuesta, tratamiento, bloque, data)
}
\arguments{
\item{respuesta}{(string) variable respuesta.}

\item{tratamiento}{(string) variable que representa los tratamientos.}

\item{bloque}{(string) variable que representan los bloques.}

\item{data}{(\code{data.frame}) Tabla de datos en formato largo con los datos.}
}
\value{
Devuelve una tabla en formato \code{data.frame} el análisis de varianza.

Devuelve una tabla en formato \code{data.frame} el análisis de varianza.
}
\description{
Realiza un calculo de anova para un diseño de bloques incompletos

Realiza un calculo de anova para un diseño de bloques incompletos
}
